Refactor relational package layout
