+++ src/fetchgraph/compat/__init__.py
@@ -1,0 +1,1 @@
+"""Compatibility shims for legacy import paths."""
