+++ src/fetchgraph/compat/json_types.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.types`."""
+from fetchgraph.relational.types import *  # noqa: F401,F403
