+++ src/fetchgraph/compat/relational_base.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.providers.base`."""
+from fetchgraph.relational.providers.base import *  # noqa: F401,F403
