+++ src/fetchgraph/compat/relational_composite.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.providers.composite_provider`."""
+from fetchgraph.relational.providers.composite_provider import *  # noqa: F401,F403
