+++ src/fetchgraph/compat/relational_models.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.models`."""
+from fetchgraph.relational.models import *  # noqa: F401,F403
