+++ src/fetchgraph/compat/relational_pandas.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.providers.pandas_provider`."""
+from fetchgraph.relational.providers.pandas_provider import *  # noqa: F401,F403
