+++ src/fetchgraph/compat/relational_schema.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.schema`."""
+from fetchgraph.relational.schema import *  # noqa: F401,F403
