+++ src/fetchgraph/compat/relational_sql.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.providers.sql_provider`."""
+from fetchgraph.relational.providers.sql_provider import *  # noqa: F401,F403
