+++ src/fetchgraph/compat/semantic_backend.py
@@ -1,0 +1,2 @@
+"""Backwards-compatible import for :mod:`fetchgraph.relational.semantic.backend`."""
+from fetchgraph.relational.semantic.backend import *  # noqa: F401,F403
