+++ src/fetchgraph/core/__init__.py
@@ -1,0 +1,52 @@
+"""Core (non-relational) components for fetchgraph."""
+
+from .context import (
+    ContextPacker,
+    BaseGraphAgent,
+    create_generic_agent,
+    make_llm_plan_generic,
+    make_llm_synth_generic,
+)
+from .models import (
+    RawLLMOutput,
+    ProviderInfo,
+    TaskProfile,
+    ContextFetchSpec,
+    BaselineSpec,
+    ContextItem,
+    RefetchDecision,
+    Plan,
+)
+from .protocols import (
+    ContextProvider,
+    SupportsFilter,
+    SupportsDescribe,
+    Verifier,
+    Saver,
+    LLMInvoke,
+)
+from .utils import load_pkg_text, render_prompt
+
+__all__ = [
+    "ContextPacker",
+    "BaseGraphAgent",
+    "create_generic_agent",
+    "make_llm_plan_generic",
+    "make_llm_synth_generic",
+    "RawLLMOutput",
+    "ProviderInfo",
+    "TaskProfile",
+    "ContextFetchSpec",
+    "BaselineSpec",
+    "ContextItem",
+    "RefetchDecision",
+    "Plan",
+    "ContextProvider",
+    "SupportsFilter",
+    "SupportsDescribe",
+    "Verifier",
+    "Saver",
+    "LLMInvoke",
+    "load_pkg_text",
+    "render_prompt",
+]
