+++ src/fetchgraph/providers/__init__.py
@@ -1,0 +1,1 @@
+"""Non-relational providers live here."""
