+++ src/fetchgraph/relational/dsl/__init__.py
@@ -1,0 +1,1 @@
+"""Relational DSL components (parsing, AST, compilation)."""
