+++ src/fetchgraph/relational/dsl/spec.yaml
@@ -1,0 +1,2 @@
+# Placeholder for relational DSL specification.
+# TODO: define keys, aliases, defaults, and operations.
