+++ src/fetchgraph/relational/providers/sql/__init__.py
@@ -1,0 +1,1 @@
+"""SQL-specific helpers for relational providers."""
