+++ src/fetchgraph/relational/semantic/__init__.py
@@ -1,0 +1,23 @@
+"""Semantic search helpers for relational providers."""
+
+from .backend import (
+    EmbeddingModel,
+    SemanticBackend,
+    CsvSemanticBackend,
+    CsvSemanticSource,
+    CsvEmbeddingBuilder,
+    PgVectorSemanticBackend,
+    PgVectorSemanticSource,
+    VectorStoreLike,
+)
+
+__all__ = [
+    "EmbeddingModel",
+    "SemanticBackend",
+    "CsvSemanticBackend",
+    "CsvSemanticSource",
+    "CsvEmbeddingBuilder",
+    "PgVectorSemanticBackend",
+    "PgVectorSemanticSource",
+    "VectorStoreLike",
+]
