+++ tests/test_relational_composite.py
@@ -4,19 +4,19 @@ import pytest
 
 pd = pytest.importorskip("pandas")
 
-from fetchgraph.relational_base import RelationalDataProvider
-from fetchgraph.relational_models import (
+from fetchgraph.relational import (
+    ComparisonFilter,
+    CompositeRelationalProvider,
     ColumnDescriptor,
     EntityDescriptor,
+    PandasRelationalDataProvider,
     QueryResult,
-    RelationalQuery,
     RelationDescriptor,
     RelationJoin,
+    RelationalDataProvider,
+    RelationalQuery,
     SelectExpr,
-    ComparisonFilter,
 )
-from fetchgraph.relational_pandas import PandasRelationalDataProvider
-from fetchgraph.relational_composite import CompositeRelationalProvider
 
 
 def _orders_provider():
