+++ tests/test_relational_composite_cross_provider.py
@@ -4,19 +4,19 @@ import pytest
 
 pd = pytest.importorskip("pandas")
 
-from fetchgraph.relational_composite import CompositeRelationalProvider
-from fetchgraph.relational_models import (
+from fetchgraph.relational import (
     AggregationSpec,
     ColumnDescriptor,
     ComparisonFilter,
+    CompositeRelationalProvider,
     EntityDescriptor,
     GroupBySpec,
+    PandasRelationalDataProvider,
     RelationalQuery,
     RelationDescriptor,
     RelationJoin,
     SelectExpr,
 )
-from fetchgraph.relational_pandas import PandasRelationalDataProvider
 
 
 def _build_block_system_composite(
