+++ tests/test_relational_sql.py
@@ -6,7 +6,7 @@ from typing import Any, List
 import pytest
 pd = pytest.importorskip("pandas")
 
-from fetchgraph.relational_models import (
+from fetchgraph.relational import (
     AggregationSpec,
     ColumnDescriptor,
     ComparisonFilter,
@@ -19,8 +19,8 @@ from fetchgraph.relational_models import (
     SelectExpr,
     SemanticClause,
     SemanticMatch,
+    SqlRelationalDataProvider,
 )
-from fetchgraph.relational_sql import SqlRelationalDataProvider
 
 
 class FakeSemanticBackend:
